moved the PlatlistApi to it's own component, added redirect file, added imports to app.js
